Removed login/logout button from footer. Added log out button.
