+++ src/App.js
@@ -1,7 +1,6 @@
 import React, {Component} from 'react';
 import {Route, BrowserRouter as Router} from "react-router-dom";
 import Home from "./containers/Home";
-import Overview from "./containers/Overview";
 import Help from "./containers/Help";
 import Charts from "./containers/Charts";
 import Footer from "./utils/Footer";
@@ -17,16 +16,11 @@ class App extends Component {
           {/* Home route */}
           <Route exact path="/" component={Home} />
 
-          {/* Overview route */}
-          <Route path="/overview" component={Overview} />
-
-          {/* Help route */}
-          <Route path="/help" component={Help} />
-
-
           {/* Charts route */}
           <Route path="/charts" component={Charts} />
 
+          {/* Help route */}
+          <Route path="/help" component={Help} />
 
           {/* Fixed footer with spacing*/}
           <Footer className="pt-10"/>
