+++ src/containers/Charts.js
@@ -1,13 +1,27 @@
 import React, { Component } from 'react';
+import connect from "react-redux/es/connect/connect";
+import {Redirect} from "react-router-dom";
+import OverviewHeader from "../utils/Header";
 
 class Charts extends Component {
-    render() {
-        return (
-            <div>
-                <h1>test</h1>
-            </div>
-        );
+  render() {
+    if (this.props.requestSent) {
+      return (
+        <div className="w-full h-screen bg-grey-lighter">
+
+          {/* Header */}
+          <OverviewHeader/>
+        </div>
+      );
     }
+    return (
+      <Redirect to='/' />
+    )
+  }
 }
 
-export default Charts;
+const mapStateToProps = (state) => ({
+  requestSent: state.requestSent,
+});
+
+export default connect(mapStateToProps)(Charts);
